+++ perf/_cli.py
@@ -154,13 +154,15 @@ def _format_runs(bench, quiet=False, verbose=False, raw=False, lines=None):
     return lines
 
 
-def _format_stats(bench, lines):
+PERCENTILE_NAMES = {0: 'minimum', 50: 'median', 100: 'maximum'}
+
+
+def format_stats(bench, lines):
     fmt = bench.format_value
     values = bench.get_values()
 
     nrun = bench.get_nrun()
     nvalue = len(values)
-    mean = bench.mean()
 
     empty_line(lines)
 
@@ -224,32 +226,49 @@ def _format_stats(bench, lines):
     lines.append('')
 
     # Minimum
-    def format_limit(mean, value):
-        return ("%s (%+.0f%% of the mean)"
-                % (fmt(value), (value - mean) * 100.0 / mean))
-
-    lines.append("Minimum: %s" % format_limit(mean, min(values)))
+    table = []
+    table.append(("Minimum", bench.format_value(min(values))))
 
     # Median +- MAD
     median = bench.median()
     if len(values) > 2:
         median_abs_dev = bench.median_abs_dev()
-        lines.append("Median +- MAD: %s +- %s"
-                     % bench.format_values((median, median_abs_dev)))
+        table.append(("Median +- MAD",
+                      "%s +- %s"
+                      % bench.format_values((median, median_abs_dev))))
     else:
-        lines.append("Mean: %s" % bench.format_value(median))
+        table.append(("Mean", bench.format_value(median)))
 
     # Mean +- std dev
     mean = bench.mean()
     if len(values) > 2:
         stdev = bench.stdev()
-        lines.append("Mean +- std dev: %s +- %s"
-                     % bench.format_values((mean, stdev)))
+        table.append(("Mean +- std dev",
+                      "%s +- %s" % bench.format_values((mean, stdev))))
     else:
-        lines.append("Mean: %s" % bench.format_value(mean))
+        table.append(("Mean", bench.format_value(mean)))
 
-    # Maximum
-    lines.append("Maximum: %s" % format_limit(mean, max(values)))
+    table.append(("Maximum", bench.format_value(max(values))))
+
+    # Render table
+    width = max(len(row[0]) + 1 for row in table)
+    for key, value in table:
+        key = (key + ':').ljust(width)
+        lines.append("%s %s" % (key, value))
+    lines.append('')
+
+    def format_limit(mean, value):
+        return ("%s (%+.0f%% of the mean)"
+                % (fmt(value), (value - mean) * 100.0 / mean))
+
+    # Percentiles
+    for p in (0, 5, 25, 50, 75, 95, 100):
+        text = format_limit(mean, bench.percentile(p))
+        text = "%3sth percentile: %s" % (p, text)
+        name = PERCENTILE_NAMES.get(p)
+        if name:
+            text = '%s -- %s' % (text, name)
+        lines.append(text)
     return lines
 
 
@@ -406,7 +425,7 @@ def format_benchmark(bench, checks=True, metadata=False,
         format_histogram([(bench, None)], lines=lines)
 
     if stats:
-        _format_stats(bench, lines=lines)
+        format_stats(bench, lines=lines)
 
     if checks:
         format_checks(bench, lines=lines)